render: support text auras: tas, ta, t
Neither `@tas` nor `@ta` explicitly sanitize, instead rely on the atom
already being "sane" for those auras. But in those cases, too, we need
to take special care: string encoding in atoms is utf-8, whereas js
strings are utf-16, so we must call a `TextDecoder`.

For `@t` we do need to sanitize the string. We lift logic previously
implemented in tloncorp/tlon-apps#3274 into the library, where it
belongs.