parse: make q implementation pass tests
We do this by making it stop doing "valid pat" checks, which are
inaccurate for the `@q` format. Instead, we catch exceptions when they
get thrown and return null for those cases.

At some point, we should more thoroughly refactor/clean up the p and q
modules, but for now we'll take a low-touch approach.