ci: remove size-limit check
As far as I can tell, this check has never run in this iteration of the
repo. Now that it is trying to run, it's failing for reasons of using
pnpm instead of npm, or some such thing. We don't care to spend much
time on it right now, so we rip the check out entirely.

We still keep it around as a script and dev dependency. For the record,
I did check the sizes from pre- and post-refactor. We have gone from
~4.7kb to ~6.6kb.