tests: improve fuzz tests
Can now generate values of arbitrary bit widths. We do this to make sure
we cover more varied inputs more consistently, and restrict fuzzing for
certain auras to inputs that make sense for that aura. (For example, no
inputs >32 bits for `@if`, because any additional bits would get
truncated, breaking the round-tripping that we test for.)

Additionally, we make it possible to fuzz floats by letting you specify
a "transform" function for the generated test values. This way, we can
ensure none of our inputs are `NaN` values, which would have their
mantissa ("payload") bits dropped during rendering.

Unfortunately, this does uncover an off-by-one in the parsing logic for
floats. In lieu of fixing that, we keep the float fuzzing disabled for
now...