parse: support UIP-135 style dates
UIP-135 expands both parsing and rendering of `@da` literals to support
leading zeroes in the month and day segments.

By accepting leading zeroes in the parser, we become compatible with
backends both before and after the UIP-135 change. Eventually, we'll
want to switch to the "fixed-width" date rendering as well, but that
doesn't make sense for this library when those backend changes haven't
shipped yet.